+++ src/types/course.ts
@@ -1,22 +1,41 @@
-export interface CourseDetails {
+export interface ICourse {
   id: number;
   name: string;
   description: string;
   startDate: string;
-  modules: Module[];
+  modules: IModule[];
 }
 
-export interface Module {
+export interface ICourses {
+  totalCourses: number;
+  pageNumber: number;
+  pageSize: number;
+  courses: ICourse[];
+}
+
+export interface IActivityType {
+  id: number;
+  name: string;
+}
+
+export interface IActivity {
   id: number;
   name: string;
   description: string;
-  activities: Activity[];
+  details: string | null;
+  typeId: number;
+  type: IActivityType;
+  startDate: string;
+  endDate: string;
+  moduleId: number;
 }
 
-export interface Activity {
+export interface IModule {
   id: number;
   name: string;
   description: string;
   startDate: string;
   endDate: string;
+  courseId: number;
+  activities: IActivity[] | [];
 }
